refactor(app_view): split Send File dialog helpers into transfer_ui.rs
Lifts the Send-File-specific UI primitives out of mod.rs:

  * `YMODEM_PROTOCOL_ID` + `transfer_protocol_opts` — the
    Select widget's default pick + option list.
  * `send_file_field_label`, `send_file_path_pill`,
    `send_file_choose_button`, `send_file_secondary_button`,
    `send_file_primary_button` — dialog-specific pill/button
    helpers. Kept separate from the generic `buttons.rs`
    primitives because they have distinct sizing + spacing
    (Send File uses a wider px_4 / py(6.0) button vs the
    generic px_3 / py_1).

Module name is `transfer_ui` rather than `transfer` to avoid
shadowing `crate::data::transfer`, which mod.rs's transfer
state-machine driving (`send_xmodem`, `send_ymodem`, etc.)
imports as `transfer::`. The `_ui` suffix accurately names the
content (Send-File-dialog UI helpers) rather than the broader
transfer concept.

The transfer **types** (`TransferIo`, `TransferState`,
`TransferResult`, `SendHexState`, `SendFileState`) stay in
mod.rs. Their fields are accessed across ~20 AppView methods
that drive the transfer state machine; moving them would force
13+ `pub(super)` field-visibility widenings for no functional
gain. The split here is "render helpers only" — same rationale
as keeping `Editor` in mod.rs while moving the form rendering.

Verified: `cargo clippy --all-targets -- -D warnings` and
`cargo test` both pass (63 tests, unchanged).
